revert measuring with react native performance
