--- app/actions/onboarding/index.ts
@@ -1,28 +1,0 @@
-import { IMetaMetricsEvent } from '../../core/Analytics';
-
-export const SAVE_EVENT = 'SAVE_EVENT';
-export const CLEAR_EVENTS = 'CLEAR_EVENTS';
-
-interface SaveEventAction {
-  type: typeof SAVE_EVENT;
-  event: IMetaMetricsEvent;
-}
-
-interface ClearEventsAction {
-  type: typeof CLEAR_EVENTS;
-}
-
-export type OnboardingActionTypes = SaveEventAction | ClearEventsAction;
-
-export function saveOnboardingEvent(event: IMetaMetricsEvent): SaveEventAction {
-  return {
-    type: SAVE_EVENT,
-    event,
-  };
-}
-
-export function clearOnboardingEvents(): ClearEventsAction {
-  return {
-    type: CLEAR_EVENTS,
-  };
-}
