+++ app/components/Views/OnboardingCarousel/index.test.tsx
@@ -1,41 +1,20 @@
 import React from 'react';
-import { waitFor } from '@testing-library/react-native';
+import { shallow } from 'enzyme';
 import OnboardingCarousel from './';
-import { NavigationProp, ParamListBase } from '@react-navigation/native';
-import { PerformanceRegressionSelectorIDs } from '../../../../e2e/selectors/PerformanceRegression.selectors';
-import renderWithProvider from '../../../util/test/renderWithProvider';
+import { Provider } from 'react-redux';
+import createMockStore from 'redux-mock-store';
 
-jest.mock('../../../util/metrics/TrackOnboarding/trackOnboarding');
-jest.mock('../../../util/test/utils', () => ({
-  isTest: true,
-}));
-
-const mockNavigate: jest.Mock = jest.fn();
-const mockSetOptions: jest.Mock = jest.fn();
-const mockNavigation = {
-  navigate: mockNavigate,
-  setOptions: mockSetOptions,
-} as unknown as NavigationProp<ParamListBase>;
-
+const mockStore = createMockStore();
+const initialState = {};
+const store = mockStore(initialState);
 
 describe('OnboardingCarousel', () => {
   it('should render correctly', () => {
-    const { toJSON } = renderWithProvider(
-      <OnboardingCarousel navigation={mockNavigation}/>
+    const wrapper = shallow(
+      <Provider store={store}>
+        <OnboardingCarousel />
+      </Provider>,
     );
-    expect(toJSON()).toMatchSnapshot();
-  });
-
-    it('should render the App Start Time text when isTest is true', async () => {
-      const { toJSON, getByTestId } = renderWithProvider(
-        <OnboardingCarousel navigation={mockNavigation}/>
-      );
-      expect(toJSON()).toMatchSnapshot();
-
-    await waitFor(() => {
-      expect(
-        getByTestId(PerformanceRegressionSelectorIDs.APP_START_TIME_ID),
-      ).toBeTruthy();
-    });
+    expect(wrapper).toMatchSnapshot();
   });
 });
