+++ app/core/Performance/Performance.test.ts
@@ -2,6 +2,10 @@
 import Performance from './Performance';
 import performance from 'react-native-performance';
 
+// Extend the global object for __DEV__
+// eslint-disable-next-line @typescript-eslint/no-shadow
+declare const global: { __DEV__: boolean };
+
 // Mock react-native-performance
 jest.mock('react-native-performance', () => {
   const originalModule = jest.requireActual('react-native-performance');
@@ -45,10 +49,9 @@ describe('Performance', () => {
     jest.resetAllMocks();
   });
 
-  it('should not log performance numbers in production', () => {
-    jest.mock('../../util/test/utils', () => ({
-      isTest: true, // or false, depending on what you want to test
-    }));
+  it('should log performance numbers in development mode', () => {
+    // Mock __DEV__ to be true to simulate development mode
+    global.__DEV__ = true;
 
     // Mock console.info to verify its calls
     console.info = jest.fn();
@@ -80,5 +83,22 @@ describe('Performance', () => {
         `APP START TIME = MAX(NATIVE LAUNCH TIME, JS BUNDLE LOAD TIME) - 1500ms`,
       ),
     );
+
+    // Clean up
+    global.__DEV__ = false; // Reset __DEV__ to its original state
+  });
+
+  it('should not log performance numbers in non-development mode', () => {
+    // Mock __DEV__ to be true to simulate development mode
+    global.__DEV__ = false;
+
+    // Set up performance service
+    Performance.setupPerformanceObservers();
+
+    // Verify that measure wasn't called
+    expect(performance.measure).not.toHaveBeenCalled();
+
+    // Verify that console.info wasn't called
+    expect(console.info).not.toHaveBeenCalled();
   });
 });
