--- e2e/selectors/PerformanceRegression.selectors.js
@@ -1,4 +1,0 @@
-
-export const PerformanceRegressionSelectorIDs = {
-  APP_START_TIME_ID: 'app-start-time-id',
-};
